requirements.txt
# Data
pandas>=2.0.0
numpy>=1.24.0
polars>=0.19.0
scipy>=1.11.0

# ML
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
statsmodels>=0.14.0

# Hyperparameter Optimization
optuna>=3.4.0

# Brokers & Exchanges
ccxt>=4.0.0
python-binance>=1.0.0
websockets>=12.0.0

# Alternative Data
yfinance>=0.2.0
requests>=2.31.0

# Web
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# Database
psycopg2-binary>=2.9.0
redis>=5.0.0
sqlalchemy>=2.0.0

# Scheduling
apscheduler>=3.10.0

# Visualization
plotly>=5.18.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
loguru>=0.7.0
PyYAML>=6.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
